@@ -1,18 +1,23 @@
 {
   "name": "theatherflix",
-  "version": "2.3.0.1 RC1.0",
   "description": "",
-  "main": "server.js",
+  "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "webpack": "webpack --progress",
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "build": "webpack --config webpack.config.js"
   },
   "author": "Nodeio Labs",
   "license": "GPL-3",
   "dependencies": {
+    "@babel/core": "^7.0.0-beta.49",
+    "@babel/preset-env": "^7.0.0-beta.49",
     "axios": "^0.18.0",
+    "babel": "^6.23.0",
     "docker": "^1.0.0",
     "express": "^4.13.4",
     "find-process": "^1.1.0",
+    "fuse-box": "^3.2.2",
     "loader": "^2.1.1",
     "material-ui": "^0.20.1",
     "modules": "^0.4.0",
@@ -24,11 +29,14 @@
     "uuid": "^3.2.1"
   },
   "devDependencies": {
+    "babel-cli": "^6.26.0",
     "babel-core": "^6.5.1",
-    "babel-loader": "^6.2.2",
-    "babel-preset-es2015": "^6.5.0",
+    "babel-loader": "^8.0.0-beta.3",
+    "babel-preset-es2015": "^6.24.1",
     "babel-preset-react": "^6.5.0",
     "babel-preset-stage-0": "^6.5.0",
-    "webpack": "^1.12.13"
+    "jsx-loader": "^0.13.2",
+    "webpack": "^4.11.1",
+    "webpack-cli": "^3.0.2"
   }
 }
