@@ -1,4 +1,10 @@
-var redux = require('redux');
+/***********************
+ * Module dependencies *
+ ***********************/
+
+ // attempt to addapt the site to redux. Left posponed. By Founder.
+ 
+ import redux from ('redux');
 
 console.log('Starting redux example');
 
