@@ -1,3 +1,6 @@
+/***********************
+ * Module dependencies *
+ ***********************/
 import React from ('react');
 
 var About = React.createClass({
@@ -8,16 +11,9 @@ var About = React.createClass({
    }
 });
 
-
+/******************
+ * Export About   *
+ ******************/
 module.exports = About;
 
-
-/* 
-Please if you code in this file please or another,  also add your email & name and date among the comments before and after any code we could write.
-Thanks a lot!
-Felipe - felipe@nodeio.us  - 13-04-2018
-Antz - cheerantz@gmail.com - 20-04-2018
-----------
-
-*/ 
 
