@@ -1,11 +1,6 @@
-/***********************
- * Module dependencies *
- ***********************/
-import redux from ('redux');
+var redux = require('redux');
 
 console.log('Starting todo redux example');
-
-// testing stuff
 
 var stateDefault = {
   searchText: '',
