@@ -5,7 +5,7 @@
 
 
 // Thatherflix is an Open Source Project
-// Founded by Felipe Alfonso Gonzalez - Soft. Engr. - f.alfonso@res-ear.ch
+// Founded by Felipe Alfonso Gonzalez - Soft. Engr. - 
 // C2018 - Present
 
 const express = require('express');
@@ -20,5 +20,3 @@ app.use(express.static('public'));
 app.listen(app.get('port'), function () {
   console.log(`Server is up, on port ${app.get('port')}`);
 });
-
-
